incremental progress toward a viable getarrivals and prefetch solutions for the API.

+++ crawl/crawler.py
@@ -6,12 +6,14 @@ import os
 from google.appengine.api import quota
 from google.appengine.api import urlfetch
 from google.appengine.api import memcache
+from google.appengine.api import datastore_errors
 from google.appengine.api.urlfetch import DownloadError
 from google.appengine.api.labs import taskqueue
 from google.appengine.api.labs.taskqueue import Task
 from google.appengine.ext import webapp
 from google.appengine.ext import db
 from google.appengine.ext.webapp import template
+from google.appengine.ext.db import GeoPt
 
 from google.appengine.runtime import apiproxy_errors
 
@@ -28,7 +30,7 @@ class CrawlerHandler(webapp.RequestHandler):
     def get(self):
         # create a new task with this link
         crawlURL = "http://webwatch.cityofmadison.com/webwatch/Ada.aspx"
-        task = Task(url='/crawlingtask', params={'crawl':crawlURL,})
+        task = Task(url='/crawlingtask', params={'crawl':crawlURL,'routeID':'00'})
         task.add('crawler')
         logging.info("Added new task for %s" % crawlURL)        
         return
@@ -40,7 +42,8 @@ class CrawlingTaskHandler(webapp.RequestHandler):
         try:
             scrapeURL = self.request.get('crawl')
             direction = self.request.get('direction')
-            logging.debug("task scraping for %s, direction %s" % (scrapeURL,direction))
+            routeID = self.request.get('routeID')
+            logging.debug("task scraping for %s, direction %s, route %s" % (scrapeURL,direction,routeID))
             
             loop = 0
             done = False
@@ -60,12 +63,12 @@ class CrawlingTaskHandler(webapp.RequestHandler):
                         time.sleep(4)
                         loop = loop+1
             end = quota.get_request_cpu_usage()
-            logging.info("scraping took %s cycles" % (end-start))
+            #logging.info("scraping took %s cycles" % (end-start))
 
             # start to interrogate the results
             soup = BeautifulSoup(result.content)
             for slot in soup.html.body.findAll("a","ada"):
-                logging.info("pulling out data from page... %s" % slot)
+                #logging.info("pulling out data from page... %s" % slot)
 
                 if slot.has_key('href'):
                     href = slot['href']
@@ -84,16 +87,19 @@ class CrawlingTaskHandler(webapp.RequestHandler):
                         logging.info("found stop %s, %s" % (stopID,intersection))
                         
                         # check for conflicts...
-                        q = db.GqlQuery("SELECT * FROM StopLocation WHERE stopID = :1", stopID)
+                        q = db.GqlQuery("SELECT * FROM StopLocation WHERE stopID = :1 and direction = :2", stopID,direction.upper())
                         stopQuery = q.fetch(1)
                         if len(stopQuery) == 0:
                             # add the new stop
                             stop = StopLocation()
                             stop.stopID = stopID
-                            stop.intersection = intersection
-                            stop.direction = direction
+                            stop.routeID = routeID
+                            stop.intersection = intersection.upper()
+                            stop.direction = direction.upper()
                             stop.put()
                             logging.info("added new stop listing MINUS geo location")
+                        else:
+                            logging.info("already have this stop in the table...")
                         
                         # pull the route and direction data from the URL
                         #routeData = scrapeURL.split('?')[1]
@@ -119,12 +125,15 @@ class CrawlingTaskHandler(webapp.RequestHandler):
                         #else:
                           #logging.error("we found a duplicate entry!?! %s", routeQuery[0].scheduleURL)
                     #else: # title.split(",")[0].isdigit():
-                    elif href.find("?r=03") > -1:
+                    elif href.find("?r=") > -1:
                         # create a new task with this link
                         crawlURL = CRAWL_URLBASE + href
-                        task = Task(url='/crawlingtask', params={'crawl':crawlURL,'direction':title})
+                        if routeID == '00' or len(routeID) > 2:
+                            routeID = href.split('r=')[1]
+                        task = Task(url='/crawlingtask', params={'crawl':crawlURL,'direction':title,'routeID':routeID})
                         task.add('crawler')
-                        logging.info("Added new task for %s, direction %s" % (title.split(",")[0],title))                    # label crawler looks for titles with letters for extraction/persistence
+                        logging.info("Added new task for %s, direction %s, route %s" % (title.split(",")[0],title,routeID))                    
+                    # label crawler looks for titles with letters for extraction/persistence
                     #elif title.replace('-','').replace(' ','').isalpha():
                     #    routeData = href.split('?')[1]
                     #    logging.info("found the route LABEL page! href: %s" % href)
@@ -178,40 +187,40 @@ class CrawlingGeoTaskHandler(webapp.RequestHandler):
             end = quota.get_request_cpu_usage()
             logging.info("scraping took %s cycles" % (end-start))
 
-            if result.content[0] == '6':
+            if result.content[0] == '7':
                 # parser for the geo stop data...
                 stops = result.content.split('<br>')
                 for s in stops:
                     #logging.info("Parsing... %s" % s)
                     if s[0] == '|':
+                        # |18;43.0656205|-89.4237496|ALLEN & COMMONWEALTH [ID#2969]|CapSq|3:20 PM TO CAP SQR
                         data = re.search('\|\d+;.*?(4.*?)\|(.*?)\|(.*?)\[ID#(\d+)\]\|(.*?)\|',s)
+                        if data is None:
+                            # |32;43.0527923|-89.43559|Monroe & Glenway|Allied-Chalet|2:40 PM TO ALLD DR:MOHAWK
+                            data = re.search('\|\d+;.*?(4.*?)\|(.*?)\|(.*?)\|(.*?)\|',s)
+                            if data is not None:
+                                latitude = data.group(1).strip()
+                                longitude = data.group(2).strip()
+                                intersection = data.group(3).strip()
+                                direction = data.group(4).strip()
+                        else:
+                            latitude = data.group(1).strip()
+                            longitude = data.group(2).strip()
+                            intersection = data.group(3).strip()
+                            stopID = data.group(4).strip()
+                            direction = data.group(5).strip()
+                                                            
                         if data is not None:
+                            logging.debug("creating task to store this stop in the datastore...")
+                            # create a task event to process the data and check/store in the datastore
+                            task = Task(url='/storethestop', params={'intersection':intersection,
+                                                                     'latitude':latitude,
+                                                                     'longitude':longitude,
+                                                                     'direction':direction,
+                                                                     'crawlLine':s,
+                                                                    })
+                            task.add('stopstorage')
                             
-                            latitude = data.group(1)
-                            longitude = data.group(2)
-                            intersection = data.group(3)
-                            stopID = data.group(4)
-                            direction = data.group(5)
-                            q = db.GqlQuery("SELECT * FROM StopLocation WHERE stopID = :1", stopID)
-                            stop = q.get()
-                            if stop is None:
-                                logging.error("We DON'T already have this stop in the store!?!")
-                            else:
-                                stop.location = (latitude+","+longitude)
-                                stop.update_location()
-                                #stop.direction = direction
-                                stop.put()
-                                #logging.info("GEO DATA ADDED for %s" % intersection)
-                                
-                                # update the route table to include a reference to the new geo data
-                                q = db.GqlQuery("SELECT * FROM RouteListing WHERE stopID = :1", stopID)
-                                route = q.get()
-                                if route is None:
-                                    logging.error("IMPOSSIBLE... no stop on record?!? %s", stopID)
-                                else:
-                                    route.stopLocation = stop
-                                    route.put()
-        
  
                 return
             
@@ -228,7 +237,7 @@ class CrawlingGeoTaskHandler(webapp.RequestHandler):
                     if title.find("[ID#") > 0:
                         # this should never happen
                         logging("FATAL ERROR")
-                    elif href.find("?r=03") > -1:
+                    elif href.find("?r=") > -1:
                         crawlURL = "http://webwatch.cityofmadison.com/webwatch/UpdateWebMap.aspx?u="+href.split("=")[1]
                         task = Task(url='/crawlinggeotask', params={'crawl':crawlURL,})
                         task.add('crawler')
@@ -244,26 +253,78 @@ class CrawlingGeoTaskHandler(webapp.RequestHandler):
 
 class DropTableHandler(webapp.RequestHandler):
     def get(self, table=""):
-        qstring = "select __key__ from " + table
+        qstring = "select * from " + table
         logging.info("query string is... %s" % qstring)
         q = db.GqlQuery(qstring)
         results = q.fetch(500)
         while results:
-            db.delete(results)
+            if table == 'RouteListing':
+                for r in results:
+                    try:
+                        sLocation = r.stopLocation
+                    except datastore_errors.Error,e:
+                        if e.args[0] == "ReferenceProperty failed to be resolved":
+                            sLocation = None
+                        else:
+                            raise
+                    if sLocation is not None:
+                        r.stopLocation = None
+                        r.put()
+            else:
+                db.delete(results)
+    
             results = q.fetch(500, len(results))
 
 ## end
     
-class StopListHandler(webapp.RequestHandler):
-    def get(self,stopID=""):
+class StopStorageHandler(webapp.RequestHandler):
+    def post(self):
+        intersection = self.request.get('intersection')
+        latitude = self.request.get('latitude')
+        longitude = self.request.get('longitude')
+        direction = self.request.get('direction')
+        logging.info("storing intersection %s at lat/lon %s,%s toward %s" % (intersection,latitude,longitude,direction))
+        
+        if len(intersection) > 400:
+            intersection = intersection.ljust(400)
+            
+        q = db.GqlQuery("SELECT * FROM StopLocation WHERE intersection = :1 AND direction = :2", intersection.upper(),direction.upper())
+        stop = q.get()
+        if stop is None:
+            logging.debug("We DON'T already have this stop in the store!?! %s to %s" % (intersection.upper(),direction.upper()))
+            # create a task event to process the error
+            task = Task(url='/crawl/errortask', params={'intersection':intersection,
+                                                        'location':(latitude+","+longitude),
+                                                        'direction':direction,
+                                                        'metaStringOne':self.request.get('crawlLine'),
+                                                        'metaStringTwo':'from geotask crawler'
+                                                      })
+            task.add('crawlerrors')
+        else:
+            stop.location = GeoPt(latitude,longitude)
+            stop.update_location()
+            #stop.direction = direction
+            stop.put()
+            logging.info("GEO DATA ADDED for %s" % intersection)
+                                
+            # update the route table to include a reference to the new geo data
+            q = db.GqlQuery("SELECT * FROM RouteListing WHERE stopID = :1", stop.stopID)
+            route = q.get()
+            if route is None:
+                logging.error("IMPOSSIBLE... no stop on record?!? %s", stop.stopID)
+            else:
+                logging.debug("updating RouteListing with this stop... %s" % stop.key())
+                route.stopLocation = stop
+                route.put()
+
         return
     
 ## end
     
 class RouteListHandler(webapp.RequestHandler):
     def get(self,routeID=""):
-      logging.info("fetching all stop locations...")
-      q = StopLocation.all()
+      logging.info("fetching all stop locations for route %s" % routeID)
+      q = db.GqlQuery("SELECT * FROM RouteListing WHERE route = :1", routeID)
       if q is not None:
           results = []
           
@@ -271,10 +332,19 @@ class RouteListHandler(webapp.RequestHandler):
           stops = q.fetch(500)
           logging.info("running through stop location list....")
           for s in stops:
+              logging.debug("fetch route list details for stop %s" % s.stopID)
+              try:
+                  sLocation = s.stopLocation
+              except datastore_errors.Error,e:
+                  if e.args[0] == "ReferenceProperty failed to be resolved":
+                      sLocation = None
+                  else:
+                      raise
+                  
               results.append({'stopID':s.stopID,
-                              'location':s.location,
-                              'intersection':s.intersection,
-                              'direction':s.direction,
+                              'location':sLocation if sLocation is not None else 'unknown',
+                              'intersection':sLocation.intersection if sLocation is not None else 'unknown',
+                              'direction':sLocation.direction if sLocation is not None else 'unknown',
                               'routeID':routeID,
                               })
               
@@ -291,11 +361,12 @@ class RouteListHandler(webapp.RequestHandler):
 
 ## end ScrapeRouteStopHandler
 def main():
-  logging.getLogger().setLevel(logging.INFO)
+  logging.getLogger().setLevel(logging.DEBUG)
   application = webapp.WSGIApplication([('/configure', CrawlerHandler),
                                         ('/configuregeo', CrawlGeoHandler),
                                         ('/crawlingtask', CrawlingTaskHandler),
                                         ('/crawlinggeotask', CrawlingGeoTaskHandler),
+                                        ('/storethestop', StopStorageHandler),
                                         ('/routelist/(.*)', RouteListHandler),
                                         ('/droptable/(.*)', DropTableHandler),
                                         ],
